# docs/swagger.yaml
definitions:
  http.FlightFullResponse:
    properties:
      description:
        type: string
      destination:
        type: string
      flight_date:
        type: string
      flight_number:
        type: string
      id:
        type: integer
      map_url:
        type: string
      origin:
        type: string
      published_at:
        type: string
      publisher_tg_id:
        type: integer
      publisher_username:
        type: string
      status:
        type: string
    type: object
  http.FlightRequest:
    properties:
      description:
        example: Лечу налегке, могу взять документы.
        type: string
      destination:
        example: Москва
        type: string
      flight_date:
        description: dd/mm/yy
        example: 10/06/25
        type: string
      origin:
        example: Санкт-Петербург
        type: string
    type: object
  http.FlightResponse:
    properties:
      flight_number:
        example: 'Рейс #1234-5678'
        type: string
      id:
        example: 17
        type: integer
    type: object
  http.OrderFullResponse:
    properties:
      cost:
        type: number
      deposit:
        type: number
      description:
        type: string
      destination_city:
        type: string
      end_date:
        type: string
      id:
        type: integer
      media_urls:
        items:
          type: string
        type: array
      order_number:
        type: string
      origin_city:
        type: string
      published_at:
        type: string
      publisher_tg_id:
        type: integer
      publisher_username:
        type: string
      reward:
        type: number
      start_date:
        type: string
      status:
        $ref: '#/definitions/order.ModerationStatus'
      store_link:
        type: string
      title:
        type: string
      type:
        $ref: '#/definitions/order.OrderType'
    type: object
  http.OrderRequest:
    properties:
      cost:
        example: 1000
        type: number
      deposit:
        example: 200
        type: number
      description:
        example: Нужно привезти из Москвы в СПб
        type: string
      destination_city:
        example: Санкт-Петербург
        type: string
      end_date:
        description: dd/mm/yy
        example: 05/06/25
        type: string
      origin_city:
        example: Москва
        type: string
      reward:
        example: 100
        type: number
      start_date:
        description: dd/mm/yy
        example: 01/06/25
        type: string
      store_link:
        example: https://store.com/item/123
        type: string
      title:
        example: Заказ на доставку
        type: string
      type:
        allOf:
        - $ref: '#/definitions/order.OrderType'
        description: <--- ЭТО ДОБАВИТЬ
        example: personal
    type: object
  http.OrderResponse:
    properties:
      id:
        type: integer
      order_number:
        type: string
    type: object
  http.TelegramAuthRequest:
    properties:
      initData:
        description: данные из телеги
        example: '...'
        type: string
    type: object
  http.TelegramAuthResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6...
        type: string
      tg_id:
        type: integer
    type: object
  order.ModerationStatus:
    enum:
    - PENDING
    - APPROVED
    - REJECTED
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusApproved
    - StatusRejected
  order.OrderType:
    enum:
    - personal
    - store
    type: string
    x-enum-varnames:
    - OrderTypePersonal
    - OrderTypeStore
info:
  contact: {}
paths:
  /auth/telegram:
    post:
      consumes:
      - application/json
      parameters:
      - description: Init data от Telegram
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/http.TelegramAuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.TelegramAuthResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Invalid signature
          schema:
            type: string
        "500":
          description: User error или Token error
          schema:
            type: string
      summary: Авторизация через Telegram
      tags:
      - auth
  /flights:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/http.FlightFullResponse'
            type: array
        "500":
          description: Ошибка при получении рейсов
          schema:
            type: string
      summary: Получить все рейсы
      tags:
      - flights
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные рейса
        in: body
        name: flight
        required: true
        schema:
          $ref: '#/definitions/http.FlightRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.FlightResponse'
        "400":
          description: Невалидный JSON или формат даты
          schema:
            type: string
        "401":
          description: Неверный или отсутствует токен
          schema:
            type: string
        "500":
          description: Ошибка сервера при создании рейса
          schema:
            type: string
      summary: Создать рейс
      tags:
      - flights
  /orders:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/http.OrderFullResponse'
            type: array
        "500":
          description: Ошибка сервера при получении заказов
          schema:
            type: string
      summary: Получить все заказы
      tags:
      - orders
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные заказа
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/http.OrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.OrderResponse'
        "400":
          description: Невалидный JSON или формат дат
          schema:
            type: string
        "401":
          description: Невалидный токен
          schema:
            type: string
        "404":
          description: Пользователь не найден
          schema:
            type: string
        "500":
          description: Ошибка сервера при создании заказа
          schema:
            type: string
      summary: Создать заказ
      tags:
      - orders
swagger: "2.0"